Changed Stock back to being a double and fixed a naming issue "AmountInStock" to stock to fit the new name

+++ src/db/sql/Create Products Table.sql
@@ -11,7 +11,7 @@ CREATE TABLE Products (
     PurchasePrice DECIMAL(10,2) NOT NULL,
     
     Location VARCHAR(50) NOT NULL,
-    Stock INT NOT NULL,
+    Stock DECIMAL(10,2) NOT NULL,
     
     Unit VARCHAR(10) NOT NULL,
